Fixing implicit 64->32 bit casts in AIUtilities.
Some of this includes throwing an assertion if a length (which can be a long on LP64) is larger than a UInt32 it's getting passed to (I'm looking at you, keychain).

+++ Frameworks/AIUtilities Framework/Source/AIArrayAdditions.m
@@ -99,7 +99,7 @@
 
 - (void)moveObject:(id)object toIndex:(unsigned)newIndex
 {
-	unsigned	currentIndex = [self indexOfObject:object];
+	NSUInteger	currentIndex = [self indexOfObject:object];
 	
 	//if we're already there, do no work
 	if (currentIndex == newIndex)
